% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{decomposition}
\alias{decomposition}
\title{Get correlation matrix decomposition}
\usage{
decomposition(fit)
}
\arguments{
\item{fit}{an efast model}
}
\value{
The decomposition returns four components: the observed correlation
matrix, the factor-implied covaraiance, the residual variance, and the
structural covariance matrix. You can plot these matrices using corrplot.
}
\description{
Get correlation matrix decomposition
}
