% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast.R
\name{efa_esem}
\alias{efa_esem}
\title{Estimate an EFA model in lavaan}
\usage{
efa_esem(data, M, auto.fix.first = FALSE, auto.var = TRUE,
  auto.efa = TRUE, information = "observed", std.ov = TRUE, ...)
}
\arguments{
\item{data}{<data.frame> the dataset}

\item{M}{<numeric> How many factors, minimum 2}

\item{auto.fix.first}{<bool> see lavaan}

\item{auto.var}{<bool> see lavaan}

\item{auto.efa}{<bool> see lavaan}

\item{information}{<character> see lavaan}

\item{std.ov}{<bool> see lavaan}

\item{...}{other arguments passed to lavaan}
}
\description{
This function estimates efa models in the same way that efast models are
estimated: using lavaan.
}
\details{
The constrained model constrains the residual covariance to be
  equal across the different ROIs.
}
\examples{
\dontrun{
# create a test dataset
test_data <- simulate_efast()
fit_efa <- efa_esem(simdat, M = 4)
summary(fit_efa)
}

}
