% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast.R
\name{efast}
\alias{efast}
\title{Estimate an EFAST model}
\usage{
efast(
  data,
  M,
  rstruct,
  auto.fix.first = FALSE,
  auto.var = TRUE,
  auto.efa = TRUE,
  information = "observed",
  std.ov = TRUE,
  ...
)
}
\arguments{
\item{data}{<data.frame> the dataset}

\item{M}{<numeric> How many factors, minimum 2}

\item{rstruct}{<list> residual structure (see details)}

\item{auto.fix.first}{<bool> see lavaan}

\item{auto.var}{<bool> see lavaan}

\item{auto.efa}{<bool> see lavaan}

\item{information}{<character> see lavaan}

\item{std.ov}{<bool> see lavaan}

\item{...}{other arguments passed to lavaan}
}
\description{
This function estimates efa models with residual covariance.
}
\details{
Residual structure (in the form of residual covariances) can be
added to the EFA through a list of pairs of variable names. See the example
for more information.
}
\examples{
\dontrun{
# Use a lavaan test dataset
test_data <- lavaan::HolzingerSwineford1939[,7:15]

# create an EFA model
test_efa <- efast(test_data, 3)

# create a (simple) residual structure
res_struct <- list(
  c("x4", "x7"),
  c("x5", "x9")
)

# create an efast model
test_efast <- efast(test_data, 3, res_struct)

compare the models
lavaan::lavTestLRT(test_efa, test_efast)
}

}
