% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast.R
\name{efast_hemi}
\alias{efast_hemi}
\title{Estimate an EFAST-hemi model}
\usage{
efast_hemi(
  data,
  M,
  lh_idx,
  rh_idx,
  roi_names,
  constrain = FALSE,
  sample.nobs = NULL,
  auto.fix.first = FALSE,
  auto.var = TRUE,
  auto.efa = TRUE,
  information = "observed",
  std.ov = TRUE,
  ...
)
}
\arguments{
\item{data}{<data.frame> the dataset or <matrix> covariance matrix}

\item{M}{<numeric> How many factors, minimum 2}

\item{lh_idx}{<numeric> column numbers of left hemisphere variables}

\item{rh_idx}{<numeric> column numbers of right hemisphere variables}

\item{roi_names}{<character> optional names of rois}

\item{constrain}{<bool> whether to constrain the symmetry (see details)}

\item{sample.nobs}{<numeric> sample size (if data = covmat, see lavaan)}

\item{auto.fix.first}{<bool> see lavaan}

\item{auto.var}{<bool> see lavaan}

\item{auto.efa}{<bool> see lavaan}

\item{information}{<character> see lavaan}

\item{std.ov}{<bool> see lavaan}

\item{...}{other arguments passed to lavaan}
}
\description{
This function estimates efast models with covariance due to hemispheric
symmetry.
}
\details{
The constrained model constrains the residual covariance to be
  equal across the different ROIs.
}
\examples{
\dontrun{
# create a test dataset
test_data <- simulate_efast()
fit_efast <- efast_hemi(simdat, M = 4, 1:17, 18:34)
summary(fit_efast)
}

}
