% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{efast_loadings}
\alias{efast_loadings}
\title{Get loadings}
\usage{
efast_loadings(fit, symmetry = FALSE)
}
\arguments{
\item{fit}{efa or efast model}

\item{symmetry}{whether to display the left and right hemisphere loadings side-by-side}
}
\description{
Get loadings
}
