% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is_efast_efa}
\alias{is_efast_efa}
\title{Check whether an object is an efast-efa model}
\usage{
is_efast_efa(x)
}
\arguments{
\item{x}{any object}
}
\description{
Check whether an object is an efast-efa model
}
