% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is_efast_hemi}
\alias{is_efast_hemi}
\title{Check whether an object is an efast-hemi model}
\usage{
is_efast_hemi(x)
}
\arguments{
\item{x}{any object}
}
\description{
Check whether an object is an efast-hemi model
}
