% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lateralization}
\alias{lateralization}
\title{Get lateralization per ROI}
\usage{
lateralization(fit)
}
\arguments{
\item{fit}{an efast model}
}
\value{
parameter estimates table of lateralization per ROI
}
\description{
The lateralization index (LI) is calculated as (sum of the uniquenesses for
the contralateral homologues) / (total residual variance after accounting for
the exploratory factors). Thus, if there is no symmetry at all, the LI is 1,
and if all the residual variance is explained by symmetry, the LI is 0.
}
