% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{efa_model}
\alias{efa_model}
\alias{str_model}
\alias{hemi_model}
\alias{lat_index}
\title{model_syntax}
\usage{
efa_model(x, M = 3)

str_model(x, rstruct)

hemi_model(rois, var_const = FALSE)

lat_index(rois)
}
\description{
model_syntax

str_model

hemi_model

lat_index stuff
}
\keyword{internal}
