% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast.R
\docType{data}
\name{roi_volume}
\alias{roi_volume}
\title{Synthesised volume data for DKT-atlas ROIs.}
\format{A data frame with 647 rows and 68 columns}
\usage{
data(roi_volume)
}
\description{
This data has been synthesised on the basis of real-world data from the
Cam-CAN cohort. It contains 68 grey matter volume measurements (34 LH and 34
RH ROIs) for 647 participants.
}
\references{
\url{cam-can.org}
}
\keyword{datasets}
