% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{simulate_efast}
\alias{simulate_efast}
\title{Simulate data from an efast matrix}
\usage{
simulate_efast(N = 650L, lam_lat = 0.595, lam_bil = 0.7,
  psi_cov = 0.5, cor_uniq = 0.4)
}
\arguments{
\item{N}{<int> Sample size}

\item{lam_lat}{<numeric> factor loading for the lateralised factor}

\item{lam_bil}{<numeric> factor loading for the bilateral factors}

\item{psi_cov}{<numeric> covariances of latent variables in (0, 1)}

\item{cor_uniq}{<numeric> residual correlation}
}
\value{
data frame with 17 regions of interest, bilaterally measured with
4 underlying factors and contralateral homology.
}
\description{
Simulate a dataset as in the simulations of Van Kesteren & Kievit (2019).
There are 17 regions of interest, measured in both the left and right
hemisphere. These ROIs have a predefined amount of correlation over and
above that expected by only the underlying factors.
}
\references{
Van Kesteren, E. J., & Kievit, R. K. (2019) Exploratory factor
    analysis with structured residuals applied to brain morphology.
}
