% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{cplot}
\alias{cplot}
\title{Create a clean correlation plot}
\usage{
cplot(mat, ...)
}
\arguments{
\item{mat}{<matrix> or <list<matrix>> the matrix or list of matrices to be
displayed.}

\item{...}{other arguments passed to corrplot::corrplot()}
}
\description{
This function is a wrapper around corrplot
}
\seealso{
\code{\link[corrplot]{corrplot}}
}
