% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast.R
\name{efa_model}
\alias{efa_model}
\alias{hemi_model}
\alias{lat_index}
\title{model_syntax}
\usage{
efa_model(dat, M = 3)

hemi_model(rois, var_const = FALSE)

lat_index(rois)
}
\description{
model_syntax

hemi_model

lat_index stuff
}
\keyword{internal}
